Add InnerExtent trait and SQLite impl
